+++ server/flags/flags.go
@@ -38,6 +38,12 @@ const (
 	EVMTracer = "evm.tracer"
 )
 
+// TLS flags
+const (
+	TLSCertPath = "tls.certificate-path"
+	TLSKeyPath  = "tls.key-path"
+)
+
 // AddTxFlags adds common flags for commands to post tx
 func AddTxFlags(cmd *cobra.Command) *cobra.Command {
 	cmd.PersistentFlags().String(flags.FlagChainID, "testnet", "Specify Chain ID for sending Tx")