+++ types/validation.go
@@ -2,6 +2,7 @@ package types
 
 import (
 	"bytes"
+	"math"
 
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/ethereum/go-ethereum/common"
@@ -26,4 +27,13 @@ func ValidateAddress(address string) error {
 		)
 	}
 	return nil
+}
+
+// SafeInt64 checks for overflows while casting a uint64 to int64 value.
+func SafeInt64(value uint64) (int64, error) {
+	if value > uint64(math.MaxInt64) {
+		return 0, sdkerrors.Wrapf(sdkerrors.ErrInvalidHeight, "uint64 value %v cannot exceed %v", value, int64(math.MaxInt64))
+	}
+
+	return int64(value), nil
 }