+++ types/validation_test.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/stretchr/testify/require"
+	"github.com/tharsis/ethermint/tests"
 )
 
 func TestIsEmptyHash(t *testing.T) {
@@ -50,5 +51,62 @@ func TestIsZeroAddress(t *testing.T) {
 
 	for _, tc := range testCases {
 		require.Equal(t, tc.expEmpty, IsZeroAddress(tc.address), tc.name)
+	}
+}
+
+func TestValidateAddress(t *testing.T) {
+	testCases := []struct {
+		name     string
+		address  string
+		expError bool
+	}{
+		{
+			"empty string", "", true,
+		},
+		{
+			"invalid address", "0x", true,
+		},
+		{
+			"zero address", common.Address{}.String(), false,
+		},
+		{
+			"valid address", tests.GenerateAddress().Hex(), false,
+		},
+	}
+
+	for _, tc := range testCases {
+		err := ValidateAddress(tc.address)
+
+		if tc.expError {
+			require.Error(t, err, tc.name)
+		} else {
+			require.NoError(t, err, tc.name)
+		}
+	}
+}
+
+func TestSafeInt64(t *testing.T) {
+	testCases := []struct {
+		name     string
+		value    uint64
+		expError bool
+	}{
+		{
+			"no overflow", 10, false,
+		},
+		{
+			"overflow", 18446744073709551615, true,
+		},
+	}
+
+	for _, tc := range testCases {
+		value, err := SafeInt64(tc.value)
+		if tc.expError {
+			require.Error(t, err, tc.name)
+			continue
+		}
+
+		require.NoError(t, err, tc.name)
+		require.Equal(t, int64(tc.value), value, tc.name)
 	}
 }